requirements.txt
faiss-cpu
langchain-community
langchain-core
langchain_huggingface
langchain-groq
python-dotenv
